python

+++ Core/Milestone 2/Python01/ex4/ft_garden_security.py
@@ -1,5 +1,5 @@
 class Plant:
-    def __init__(self, name: str, height: int, age: int):
+    def __init__(self, name: str, height: float, age: int):
         self.name = name
         self._height = height
         self._age = age
@@ -14,7 +14,7 @@ class Plant:
     def aged(self):
         self._age += 1
 
-    def set_height(self, height: int):
+    def set_height(self, height: float):
         if (height < 0):
             print(f'{self.name}: Error, height can\'t be negative')
             print("Height update rejected")
@@ -30,7 +30,7 @@ class Plant:
             self._age = age
             print(f'Age updated: {self.get_age()} days')
 
-    def get_height(self) -> int:
+    def get_height(self) -> float:
         return (self._height)
 
     def get_age(self) -> int:
